apply the same changes to functor search
