+++ src/routes/functor-search/+page.svelte
@@ -1,22 +1,23 @@
 <script lang="ts">
 	import { goto } from '$app/navigation'
-	import { fade } from 'svelte/transition'
-	import { browser } from '$app/environment'
 	import Selection from '$components/Selection.svelte'
 	import { encode_property_ID } from '$lib/commons/property.url'
 	import MetaData from '$components/MetaData.svelte'
 	import { SEARCH_SEPARATOR } from '$lib/commons/search.config'
-	import { pluralize } from '$lib/client/utils'
-	import FunctorList from '$components/FunctorList.svelte'
+	import { navigating } from '$app/state'
 
 	let { data } = $props()
 
-	// svelte-ignore state_referenced_locally
-	let satisfied_properties = $state(data.satisfied_properties)
-	// svelte-ignore state_referenced_locally
-	let unsatisfied_properties = $state(data.unsatisfied_properties)
+	let satisfied_properties: string[] = $state([])
+	let unsatisfied_properties: string[] = $state([])
+
+	let is_valid_search = $derived(
+		satisfied_properties.length > 0 || unsatisfied_properties.length > 0,
+	)
 
 	function request_search_results() {
+		if (!is_valid_search) return
+
 		const satisfied_query = satisfied_properties
 			.map(encode_property_ID)
 			.join(SEARCH_SEPARATOR)
@@ -25,34 +26,17 @@
 			.map(encode_property_ID)
 			.join(SEARCH_SEPARATOR)
 
-		const url = new URL('/functor-search', window.location.origin)
+		const url = new URL('/functor-search/results', window.location.origin)
 
 		if (satisfied_query) url.searchParams.set('satisfied', satisfied_query)
 		if (unsatisfied_query) url.searchParams.set('unsatisfied', unsatisfied_query)
 
-		goto(url, { invalidateAll: true })
+		goto(url)
 	}
 
-	function get_dual_search_results_link() {
-		if (!data.dual_search_available || !browser) return '/'
+	const sample_search_url = `/functor-search/results?satisfied=continuous&unsatisfied=cocontinuous`
 
-		const satisfied_query = (data.dual_satisfied_properties as string[])
-			.map(encode_property_ID)
-			.join(SEARCH_SEPARATOR)
-
-		const unsatisfied_query = (data.dual_unsatisfied_properties as string[])
-			.map(encode_property_ID)
-			.join(SEARCH_SEPARATOR)
-
-		const url = new URL('/functor-search', window.location.origin)
-
-		if (satisfied_query) url.searchParams.set('satisfied', satisfied_query)
-		if (unsatisfied_query) url.searchParams.set('unsatisfied', unsatisfied_query)
-
-		return decodeURIComponent(url.toString())
-	}
-
-	const sample_search_url = `/functor-search?satisfied=continuous&unsatisfied=cocontinuous`
+	let is_searching = $derived(navigating.to?.route.id === '/functor-search/results')
 </script>
 
 <MetaData
@@ -64,55 +48,35 @@
 
 <p class="hint">
 	Search for functors with certain properties while excluding others. For example, you
-	can <a href={sample_search_url} data-sveltekit-reload="true">look</a>
+	can <a href={sample_search_url}>look</a>
 	for functors that are continuous but not cocontinuous.
 </p>
 
-<div class="form">
-	<Selection
-		title="Looking for functors with these properties:"
-		bind:selected_items={satisfied_properties}
-		allowed_items={data.all_properties}
-		section_label="Satisfied properties"
-		item_label="Satisfied property"
-	/>
+<Selection
+	title="Looking for functors with these properties:"
+	bind:selected_items={satisfied_properties}
+	allowed_items={data.all_properties}
+	section_label="Satisfied properties"
+	item_label="Satisfied property"
+/>
 
-	<Selection
-		title="... and <i>not</i> with these properties:"
-		bind:selected_items={unsatisfied_properties}
-		allowed_items={data.all_properties}
-		section_label="Unsatisfied properties"
-		item_label="Unsatisfied property"
-	/>
+<Selection
+	title="... and <i>not</i> with these properties:"
+	bind:selected_items={unsatisfied_properties}
+	allowed_items={data.all_properties}
+	section_label="Unsatisfied properties"
+	item_label="Unsatisfied property"
+/>
 
-	<button type="button" class="button" onclick={request_search_results}>Search</button>
-</div>
-
-{#if data.is_search}
-	<section transition:fade={{ duration: 150 }}>
-		<h2>Results</h2>
-
-		<p class="hint">
-			{pluralize(data.found_functors.length, {
-				one: 'Found {count} functor',
-				other: 'Found {count} functors',
-			})}
-		</p>
-		<FunctorList functors={data.found_functors ?? []} />
-
-		{#if data.dual_search_available}
-			<p class="hint">
-				All selected properties have a dual, you may perform the
-				<a href={get_dual_search_results_link()} data-sveltekit-reload="true">
-					dual search
-				</a>.
-			</p>
-		{/if}
-	</section>
-{/if}
-
-<style>
-	.form {
-		margin-bottom: 2rem;
-	}
-</style>
+<button
+	type="button"
+	class="button"
+	onclick={request_search_results}
+	disabled={is_searching || !is_valid_search}
+>
+	{#if is_searching}
+		Searching...
+	{:else}
+		Search
+	{/if}
+</button>
